requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13
openai==1.12.0
python-dotenv==1.0.0
pydub==0.25.1
wikipedia==1.4.0
requests==2.31.0
beautifulsoup4==4.12.2
pytz==2023.3
gTTS==2.3.2
playsound==1.3.0
psutil==5.9.5
pywin32==306 